lazy: column definitions 🍺
