+++ client/i18n/nn.json
@@ -829,7 +829,8 @@
     "autoMatchScoreHint": "Du har automatisk matchet **{{autoMatch}}%** av de totale tidsoppføringene, noe som betyr at **{{manualMatch}}%** av oppføringene har blitt matchet manuelt i applikasjonen.",
     "autoMatchError": "Den automatiske matchingsscoren er kritisk lav (under 70 %). For bedre resultater bør du kategorisere Outlook-avtalene dine og inkludere korrekt prosjekttagg i emnefeltet eller brødteksten.",
     "autoMatchWarning": "Den automatiske matchingsscoren er under 85 %. For å forbedre nøyaktigheten, bør du sørge for at Outlook-avtalene dine er riktig kategorisert, og gjerne legge til prosjekttagg i emnefeltet eller brødteksten til avtalene.",
-    "autoMatchSuccess": "Den automatiske matchingsscoren er veldig høy (over 95 %). Outlook-avtalene dine ser ut til å være godt strukturert med riktige kategorier og prosjekttagger. Fortsett med dette for å sikre presis registrering."
+    "autoMatchSuccess": "Den automatiske matchingsscoren er veldig høy (over 95 %). Outlook-avtalene dine ser ut til å være godt strukturert med riktige kategorier og prosjekttagger. Fortsett med dette for å sikre presis registrering.",
+    "periodLabel": "Periode"
   },
   "navigation": {
     "TimesheetPage": "Timeliste",
